sadsd

+++ scripts/vextank.py
@@ -84,10 +84,10 @@ def scan_publisher():
         us.publish(UInt16(range))
         rospy.loginfo("SCAN:" + str(range))
 
-        left_encoder = rospy.Publisher('lwheel', int, queue_size=10)
-        right_encoder = rospy.Publisher('rwheel', int, queue_size=10)
-        left_encoder.publish(int(BrickPi.Encoder[LF_WHEEL]))
-        right_encoder.publish(int(BrickPi.Encoder[RF_WHEEL]))
+        left_encoder = rospy.Publisher('lwheel', Int64, queue_size=10)
+        right_encoder = rospy.Publisher('rwheel', Int64, queue_size=10)
+        left_encoder.publish(BrickPi.Encoder[LF_WHEEL])
+        right_encoder.publish(BrickPi.Encoder[RF_WHEEL])
 
 if __name__ == '__main__':
     try:
